+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Latest
+## 3.2.0 (2019-03-18)
 
 - [Improvement] `images pull` now also pulls vendor images
 - [Feature] Add convenient `config printvalue` command